Fix no-cgo build
